Payments Double -> BigDecimal

+++ src/main/java/com/example/cmtProject/entity/erp/salaries/Payment.java
@@ -1,18 +1,15 @@
 package com.example.cmtProject.entity.erp.salaries;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.example.cmtProject.dto.erp.salaries.PaymentDTO;
-import com.example.cmtProject.dto.erp.salaries.SalaryItemDTO;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -44,31 +41,31 @@ public class Payment { // 급여 지급 이력 엔티티
     private Long payBasic; // 기본급
 
     @Column(name = "PAY_BONUS_OVERTIME")
-    private Double payBonusOvertime; // 야근수당
+    private BigDecimal payBonusOvertime; // 야근수당
 
     @Column(name = "PAY_BONUS_HOLIDAY")
-    private Double payBonusHoliday; // 명절수당
+    private BigDecimal payBonusHoliday; // 명절수당
 
     @Column(name = "PAY_BONUS_TOTAL", nullable = false)
     private Long payBonusTotal; // 총수당금액
 
     @Column(name = "PAY_TAX_PENSION")
-    private Double payTaxPension; // 국민연금
+    private BigDecimal payTaxPension; // 국민연금
 
     @Column(name = "PAY_TAX_CARE")
-    private Double payTaxCare; // 장기요양보험
+    private BigDecimal payTaxCare; // 장기요양보험
 
     @Column(name = "PAY_TAX_HEALTH")
-    private Double payTaxHealth; // 건강보험
+    private BigDecimal payTaxHealth; // 건강보험
 
     @Column(name = "PAY_TAX_EMPLOYMENT")
-    private Double payTaxEmployment; // 고용보험
+    private BigDecimal payTaxEmployment; // 고용보험
 
     @Column(name = "PAY_TAX_INCOME")
-    private Double payTaxIncome; // 소득세
+    private BigDecimal payTaxIncome; // 소득세
 
     @Column(name = "PAY_TAX_RESIDENCE")
-    private Double payTaxResidence; // 주민세
+    private BigDecimal payTaxResidence; // 주민세
 
     @Column(name = "PAY_TAX_TOTAL", nullable = false)
     private Long payTaxTotal; // 총공제금액
